Move mouse handler to its own module
